Move filter button to a separate component
