+++ frontend/app/group/[id]/index.tsx
@@ -1,5 +1,4 @@
 import Header, { HEADER_HEIGHT } from '@components/Header'
-import { Icon } from '@components/Icon'
 import { Pane, PaneHeader } from '@components/Pane'
 import { RoundIconButton } from '@components/RoundIconButton'
 import { Text } from '@components/Text'
@@ -8,105 +7,22 @@ import { GroupSplitsList } from '@components/groupScreen/GroupSplitsList'
 import { MembersButton } from '@components/groupScreen/MembersButton'
 import { MembersList } from '@components/groupScreen/MembersList'
 import { MembersOrderFilter } from '@components/groupScreen/MembersOrderFilter'
+import { SplitQueryButton } from '@components/groupScreen/SplitQueryButton'
 import { useGroupInfo } from '@hooks/database/useGroupInfo'
 import { useGroupPermissions } from '@hooks/database/useGroupPermissions'
-import { resetSplitQueryConfig, useSplitQueryConfig } from '@hooks/useSplitQueryConfig'
 import { useTheme } from '@styling/theme'
 import { DisplayClass, useDisplayClass, useThreeBarLayout } from '@utils/dimensionUtils'
-import { defaultQueryConfig } from '@utils/splitQueryConfig'
-import { useLocalSearchParams, useRouter } from 'expo-router'
+import { useLocalSearchParams } from 'expo-router'
 import React from 'react'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Platform, Pressable, ScrollView, StyleSheet, View } from 'react-native'
-import { Gesture, GestureDetector } from 'react-native-gesture-handler'
 import Animated, {
   FadeIn,
   FadeOut,
-  LinearTransition,
-  useAnimatedStyle,
-  withTiming,
 } from 'react-native-reanimated'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { GroupUserInfo } from 'shared'
-
-function SplitQuery() {
-  const theme = useTheme()
-  const router = useRouter()
-  const { id } = useLocalSearchParams()
-  const groupId = Number(id)
-  const query = useSplitQueryConfig(groupId)
-  const [pressed, setPressed] = useState(false)
-
-  const queryApplied = query !== defaultQueryConfig
-
-  const longPress = Gesture.LongPress()
-    .runOnJS(true)
-    .onStart(() => {
-      resetSplitQueryConfig(groupId)
-    })
-
-  const backgroundStyle = useAnimatedStyle(() => {
-    return {
-      backgroundColor: withTiming(queryApplied ? theme.colors.primary : 'transparent', {
-        duration: 250,
-      }),
-    }
-  })
-
-  const innerBackgroundStyle = useAnimatedStyle(() => {
-    return {
-      backgroundColor: withTiming(pressed ? `${theme.colors.onPrimary}40` : 'transparent', {
-        duration: 250,
-      }),
-    }
-  })
-
-  return (
-    <GestureDetector gesture={longPress}>
-      <Animated.View
-        layout={Platform.OS !== 'web' ? LinearTransition : undefined}
-        style={[backgroundStyle, { borderRadius: 24, overflow: 'hidden' }]}
-      >
-        <Animated.View
-          layout={Platform.OS !== 'web' ? LinearTransition : undefined}
-          style={innerBackgroundStyle}
-        >
-          <Pressable
-            disabled={!queryApplied}
-            onPressIn={() => setPressed(true)}
-            onPressOut={() => setPressed(false)}
-            onPress={() => router.navigate(`/group/${groupId}/filter`)}
-            style={{
-              paddingHorizontal: queryApplied ? 4 : 0,
-              flexDirection: 'row',
-              justifyContent: 'center',
-              alignItems: 'center',
-              gap: 4,
-            }}
-          >
-            {queryApplied && (
-              <Icon
-                name='check'
-                size={24}
-                color={theme.colors.onPrimary}
-                style={{ marginLeft: 8 }}
-              />
-            )}
-            <Animated.View layout={Platform.OS !== 'web' ? LinearTransition : undefined}>
-              <RoundIconButton
-                icon='filter'
-                disabled={queryApplied}
-                color={queryApplied ? theme.colors.onPrimary : undefined}
-                onPress={() => router.navigate(`/group/${groupId}/filter`)}
-              />
-            </Animated.View>
-          </Pressable>
-        </Animated.View>
-      </Animated.View>
-    </GestureDetector>
-  )
-}
 
 function SingleColumnLayout({ info }: { info: GroupUserInfo | undefined }) {
   const theme = useTheme()
@@ -139,7 +55,7 @@ function SingleColumnLayout({ info }: { info: GroupUserInfo | undefined }) {
               title={t('tabs.splits')}
               textLocation='start'
               adjustsFontSizeToFit
-              rightComponent={permissions?.canQuerySplits() && <SplitQuery />}
+              rightComponent={permissions?.canQuerySplits() && <SplitQueryButton />}
             />
           </View>
         </View>
@@ -198,7 +114,7 @@ function TripleColumnLayout({ groupInfo }: { groupInfo: GroupUserInfo | undefine
           containerStyle={{ flex: 1 }}
           collapsible
           collapsed={false}
-          rightComponent={permissions?.canQuerySplits() && <SplitQuery />}
+          rightComponent={permissions?.canQuerySplits() && <SplitQueryButton />}
         >
           <GroupSplitsList info={groupInfo} />
         </Pane>
